+++ src/Category_Colors/Main.php
@@ -15,7 +15,7 @@ class Main {
 	public static $version;
 
 	public $text_colors = array(
-		'Default' => 'inherit',
+		'Default' => 'no_color',
 		'Black'   => '#000',
 		'White'   => '#fff',
 		'Gray'    => '#999',